Disabled-ADComputers Problem Resolved
Finished working out the bugs in Disabled-ADComputer functions.

Want to make Delete-ADComputer more similar to the way
Disable-ADComputer is laid out.

Still need to work out a way to grab processed data and then Export it
into a SQL Table and send an HTML Email with a summary of the processed
Data.