Cleaning up the AD Comptuer Cleanup Script
Worked on cleaning up the ADComputer Cleanup Script to make it more
functional again.

Disable-ADComputers function needs a complete overhal because passing
variables to though the pipeline to WorkerDisableComputer is not working
correctly.

Still need to built in saving to a database support, and need a check to
see if the folders it is trying save the csv files to are there.

Currently the script works but it is not saving the results of the
disabled computers that it move to either the csv or in the email.  I
still think that the script could be a lot cleaner.