Pull Information from Website and Rename / Move Files
Pulls information from an HTML URL and then compares local files in a
folder that you specify.  First it looks for chapters and then creates a
folder per chapter if one doesn't already exist.  Then it compares mp4
files in the folder path you specified and if it finds a match renames
it to the title it pulled from the website.  After that its finished it
move the mp4 files to their correct chapter folder.