Version 1.1
No longer changes directory when executing the script which was really
annoying.  Script now looks at for with starting from the back to the
front just in case there is one in the course title, which does happen
and was causing an error.

Also configured the script to not look at substring when trying leading
characters because count was changing and causing an error.  Now it
looks for special characters to do the split on which is what it should
have been doing since the beginning.