Version 1.2
Updated the splits to that it now can handle dashes in the course name
and not get hung up on those.  Added a check to see if the author check
failed and exit before getting error messages on invoke-WebRequest and
causing the script to hang.  Small changes to verbose & help.