Version 1.3
Issue found with the variables not working correctly when sending the
account disabled email.

Script now compares the day that the password expires and the current
day, no longer looks at the time.

Added a DisableExpiredAccount parameter so that the number of days
before the account should be disabled can be specified at runtime of the
script.

Changed several places variable character case output.