Version 1.4
Updated the way that it displays the domain of where the script is being
ran from.  Configured the list users account information included in the
email to have the subject as bold and to also added the domain of the
user in the list of information provided.  Changed the wording in
several places so that it made more since and provided more information.