Version 1.6
Added Manager and Department as well as removed description in the list
of user account information.  Updated Email Subjects text. Updated Email
text in the body of the emails so that it doesn't say that the password
has been expired for 0 days. If the manager property is specified in the
AD Account, their manager will be CCed on the disabled account email.