Version 1.2 Get-LockedADUsersZendesk.ps1
Added ZendeskCollaborators parameter so that email address could be
specified on the CC filed with the tickets are being created.  Now DC,
Department, Manager and Manager Email Address are now utilized from the
Get-LockedADUsers Function.  Fixed some typos as well as corrected the
effort field to only read 0-30_minutes instead of 30-60_minutes.